// apps/backend/src/ecourts/interfaces/ecourts.interface.ts
export interface Court {
  id: string;
  name: string;
  state: string;
  district: string;
  type: CourtType;
  address: string;
  phone?: string;
  email?: string;
  website?: string;
}

export enum CourtType {
  SUPREME_COURT = 'SUPREME_COURT',
  HIGH_COURT = 'HIGH_COURT',
  DISTRICT_COURT = 'DISTRICT_COURT',
  SESSIONS_COURT = 'SESSIONS_COURT',
  CIVIL_COURT = 'CIVIL_COURT',
  FAMILY_COURT = 'FAMILY_COURT',
  CONSUMER_COURT = 'CONSUMER_COURT',
  LABOUR_COURT = 'LABOUR_COURT',
  REVENUE_COURT = 'REVENUE_COURT',
  OTHER = 'OTHER',
}
// ...
export interface ScrapingResult<T> {
  data: T;
  source: string;
  scrapedAt: string;
  cacheKey: string;
}